Improve safety with rollback when modifying modules actions
The code block modifies the script modules actions in order to provide
an opportunity to dequeue a script module that is not necessary.

This change bails and may restore removed actions in case there's a
problem removing the actions.

This would result in an empty script module being enqueued on the page.
It would be harmless other than adding an unnecessary request.